requirements.txt
astropy
numpy
pyyaml
scipy
mpmath